+++ src/transactions/VerifiableTransaction.d.ts
@@ -20,9 +20,9 @@ export declare interface SignedTransaction {
 }
 
 export declare class VerifiableTransaction {
-    static createTransactionHash(transactionPayload: string): string
+    static createTransactionHash(transactionPayload: string, generationHash: string): string
 
-    public signTransaction(keyPair: any): {payload: string, hash:string}
+    public signTransaction(keyPair: any, generationHash: string): {payload: string, hash:string}
 
     public signCosignatoriesTransaction(keyPair: any): {parentHash: string, signature:string, signer: string}